Updated to use full import paths rather than relative paths and moved fixtures to conftest.
